General shaders changes.
Added a new gubo for the point light model and used it for the Tour Eiffel. Updated some ubos to have GGX parameters to use the same shader for every monument. Pagoda's shaders are a placeholder -> the poi light will be added to every monument. Updated old shaders to use the color passed with the gubo.